Remove warnings from build.
The build generates several warnings due to unused subroutines
and variables.
toplevel_cpp.mk adds a config -DREMOVE_WARNINGS
which is used to remove the cause of warnings in
multiple files.

is used on multiple files.

is used once.

Otherwise dbg_log is called to use return values.
This will add some small amount of code, but not
significantly so

Signed-off-by: Ulf Samuelsson <ulf.samuelsson@telia.com>